Short by default

+++ VoxelStatsWrite.m
@@ -1,7 +1,7 @@
 function VoxelStatsWrite( data, filename, ref_file, noOfSlices, imageType )
 switch nargin
     case 4
-        imageType = 'float';
+        imageType = 'short';
 end
 h = newimage(filename, [0 noOfSlices], ref_file, imageType);
 putimages(h, data, 1:noOfSlices);
